data/sequence buffers

+++ src/main/java/io/github/thepun/pq/Persister.java
@@ -273,7 +273,7 @@ final class Persister implements Runnable {
                 sequence.next();
                 sequence.setId(nextSequenceId);
                 sequence.setElementCursor(initialDataCursor);
-                sequence.setElementLength(writer.getCursor() - initialDataCursor);
+                sequence.setElementLength((int)(writer.getCursor() - initialDataCursor));
                 sequence.setElementType(marshallerIds[typeHash]);
                 sequence.commit(nextSequenceId);
                 nextSequenceId++;
